requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
PyPDF2

langchain
langchain-core
langchain-community
langchain-pinecone
langchain-google-genai

pinecone
google.generativeai
jinja2